+++ bluefin-examples/CHANGELOG.md
@@ -1,5 +1,7 @@
 * Add `Bluefin.Examples.Pipes` and `Bluefin.Examples.Pipes.Prelude`
 
+* Add `Bluefin.Examples.Prim`
+
 ## 0.0.0.0
 
 * Implement DB example/template.  Thanks to Florian Beeres @cideM.