+++ bluefin/CHANGELOG.md
@@ -1,5 +1,7 @@
 * Add `Bluefin.DslBuilderEff`
 
+* Add `Bluefin.Prim`
+
 # 0.4.1.0
 
 * Depend on `bluefin-internal >= 0.4.1.0` to pick up `MonadFix`